v1.4.8 ~同步1.7.6版本.net内核 ~同步最新版本GeoIP数据库文件 ~在线生成配置文件时，默认添加udp支持 ~修复个人sub源无法使用的bug ~修复小闪存模式下面板无法正常更新的bug

+++ scripts/start.sh
@@ -638,8 +638,7 @@ bfstart(){
 		fi
 	fi
 	#检查dashboard文件
-	if [ -f $clashdir/ui/index.html ];then
-		rm -rf $bindir/ui
+	if [ -f $clashdir/ui/index.html -a ! -f $bindir/ui/index.html ];then
 		cp -rf $clashdir/ui $bindir
 	fi
 	#检查curl或wget支持
